move Stage to common
