+++ ebl/corpus/domain/chapter.py
@@ -14,7 +14,7 @@ from ebl.errors import NotFoundError
 from ebl.merger import Merger
 from ebl.transliteration.domain.markup import MarkupPart, to_title
 from ebl.transliteration.domain.museum_number import MuseumNumber
-from ebl.transliteration.domain.stage import Stage
+from ebl.common.domain.stage import Stage
 from ebl.transliteration.domain.text_id import TextId
 from ebl.transliteration.domain.text_line import TextLine
 from ebl.transliteration.domain.translation_line import (
