+++ ebl/corpus/domain/text.py
@@ -4,7 +4,7 @@ import attr
 
 from ebl.bibliography.domain.reference import Reference
 from ebl.corpus.domain.chapter import make_title
-from ebl.transliteration.domain.stage import Stage
+from ebl.common.domain.stage import Stage
 from ebl.transliteration.domain.text_id import TextId
 from ebl.transliteration.domain.genre import Genre
 from ebl.transliteration.domain.markup import MarkupPart
