+++ ebl/tests/corpus/test_corpus_text.py
@@ -1,5 +1,5 @@
 from ebl.corpus.domain.chapter import Classification, make_title
-from ebl.transliteration.domain.stage import Stage
+from ebl.common.domain.stage import Stage
 from ebl.corpus.domain.text import ChapterListing, Text, UncertainFragment
 from ebl.transliteration.domain.text_id import TextId
 from ebl.transliteration.domain.museum_number import MuseumNumber
