+++ ebl/tests/factories/ids.py
@@ -2,7 +2,7 @@ import factory
 from ebl.transliteration.domain.genre import Genre
 
 from ebl.transliteration.domain.parallel_line import ChapterName
-from ebl.transliteration.domain.stage import Stage
+from ebl.common.domain.stage import Stage
 from ebl.transliteration.domain.text_id import TextId
